Refactor distribution to distribution module
